+++ client.go
@@ -36,6 +36,10 @@ type Client struct {
 	Upgrade         UpgradeService
 	Features        FeatureService
 	Vpcs            VpcService
+	Drives          DriveService
+	Network         NetworkService
+	Egressgateways  EgressgatewayService
+	Egressips       EgressipService
 }
 
 // DefaultClientOptions read from the environment (BL_API_KEY, BL_CLIENT_CREDENTIALS,
@@ -90,6 +94,10 @@ func NewClient(opts ...option.RequestOption) (r Client) {
 	r.Upgrade = NewUpgradeService(opts...)
 	r.Features = NewFeatureService(opts...)
 	r.Vpcs = NewVpcService(opts...)
+	r.Drives = NewDriveService(opts...)
+	r.Network = NewNetworkService(opts...)
+	r.Egressgateways = NewEgressgatewayService(opts...)
+	r.Egressips = NewEgressipService(opts...)
 
 	return
 }